requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cx-Oracle==8.1.0
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-bootstrap4==2.3.1
django-currentuser==0.5.3
django-filter==2.4.0
django-precise-bbcode==1.2.13
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
filelock==3.0.12
flake8==3.8.4
idna==2.10
isort==5.8.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
oauthlib==3.1.0
parso==0.7.1
Pillow==8.1.2
pluggy==0.13.1
protobuf==3.15.7
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==6.0.0
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.7.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
six==1.15.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.3
wrapt==1.12.1
yapf==0.31.0
